baocao_thutuc

+++ database/migrations/2021_04_12_074034_create_thutuc_table.php
@@ -18,8 +18,8 @@ class CreateThutucTable extends Migration
             $table->string('maTT');
             $table->string('tenTT');
             $table->tinyInteger('danhmuc_id');
-            $table->text('mota') ->nullable();
-            $table->integer('tg_giai_quyet');
+            $table->text('mota')->nullable();
+            $table->tinyInteger('tg_giai_quyet');
             $table->timestamps();
         });
     }
